[HigherOrderOp] Fix returning captured value
Fixes #104298.

The bug was:
- we were only checking for freevars in SubgraphTracer.create_proxy
- freevars can also show up in SubgraphTracer.create_node

This PR adds handles free variable handling of the output of the graph
(which is created via `create_node`) in `speculate_subgraph`.

Because `create_proxy` calls `create_node`, you may be wondering why we
can't do the freevar lifting in `create_node`. The answer is that:
- `create_node` only gets used by Dynamo to create outputs of a graph,
so it is called rarely. All other callsites go through `create_proxy`.
- our freevar system is based off of VariableTrackers being associated with
Proxy objects which are associated with a SubgraphTracer.
- `create_proxy` accepts Proxy args while `create_node` accepts Node args
- Given a node, there isn't a way to retrieve the existing proxy that
wraps the node.

Test Plan:
- add new tests

ghstack-source-id: 40c53dc67e2fceae63eac665bebb15baed46d394
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104371