[HigherOrderOps] Clean up side effect handling
I think after https://github.com/pytorch/pytorch/pull/104077, we don't
need to do a diff between the SideEffects object before/after for
HigherOrderOps -- the ability is baked into speculate_subgraph.

The rationale for this PR is that diff-ing the SideEffects object didn't
work very well: it was overly conservative. If a variable gets
tracked for mutation, or a new cell variable is created, then the
SideEffects object changes.

The SideEffects object tracks two types of side effects:
- variable assignment/modification. This is covered by
[check_allowed_side_effect](https://github.com/pytorch/pytorch/blob/main/torch/_dynamo/side_effects.py#L146C9-L146C34)
- save_for_backward tracking. I don't think we even need to track this;
if the inputs require grad, then we cannot graph break in the middle of
autograd.Function, so we never need to replay calling `save_for_backward`.
If the inputs don't require grad, then `save_for_backward` doesn't do
anything, so it doesn't need to be replayed either. If we wanted to be
safe we could also call `check_allowed_side_effect` there.

Test Plan:
- #104077 introduced some heavy testing already. This PR adds some more
test cases.

ghstack-source-id: cf51d632b79cb2e7d048b08153d6891aa330e3e7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104685